requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.11

backoff==2.2.1
DateTime==5.5